Docker image export
this isn't ready quite yet but needs to be added now.
